Address code scanning issues
